LayoutTests/accessibility/svg-text.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<svg xmlns="http://www.w3.org/2000/svg">
    <text id="text1">Hello World!</text>
</svg>

<div id="console"></div>

<script>
    description("This tests SVGText elements and their content are properly handled as static text.");

    if (window.accessibilityController) {
        text1 = accessibilityController.accessibleElementById("text1");
        shouldBe("text1.role", "'AXRole: AXGroup'");
        shouldBe("text1.childrenCount", "1");
        staticText = text1.childAtIndex(0);
        shouldBe("staticText.role", "'AXRole: AXStaticText'");
        shouldBe("staticText.stringValue", "'AXValue: Hello World!'");
    }
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
